+++ CHANGELOG.md
@@ -14,6 +14,11 @@
   changed.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#2249](https://github.com/realm/SwiftLint/issues/2249)
+  
+  * Add `allow_first_closure_on_same_line` configuration to 
+  `multiline_arguments`  
+  [Mike Ciesielka](https://github.com/maciesielka)
+  [#1896](https://github.com/realm/SwiftLint/issues/1896)
 
 #### Bug Fixes